pyproject.toml
[project]
name = "web-scraping-interview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "fastapi[standard]>=0.115.6",
    "google-cloud-bigquery>=3.27.0",
    "jsonpath-ng>=1.7.0",
    "pandas-gbq>=0.26.1",
    "pandas>=2.2.3",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
